Default Visibility setting help text suggestion
Suggested correction of help text for Default Visibility setting. 

The original help text is: “If it is turned on then all books will be visible to everyone.”

A literal reading of this help text might lead one to believe that ticking this box and clicking the Save changes button would result in all the books contained within the current Booktype instance having their visibility settings changed to public. This obviously isn't the case (I hope!). Maybe the proposed text is more reassuring.